refactor(field-help): convert files to typescript
