--- src/__internal__/field-help/index.d.ts
@@ -1,1 +1,0 @@
-export { default } from "./field-help";
