--- src/__internal__/field-help/index.js
@@ -1,1 +1,0 @@
-export { default } from "./field-help.component";
