+++ src/__internal__/field-help/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./field-help.component";
+export type { FieldHelpProps } from "./field-help.component";
